Rename module iota to client
