+++ README.md
@@ -88,8 +88,8 @@ use identity_iota::account::IdentitySetup;
 use identity_iota::account::Result;
 use identity_iota::account_storage::Stronghold;
 use identity_iota::core::ToJson;
-use identity_iota::iota::ExplorerUrl;
-use identity_iota::iota::ResolvedIotaDocument;
+use identity_iota::client::ExplorerUrl;
+use identity_iota::client::ResolvedIotaDocument;
 
 #[tokio::main]
 async fn main() -> Result<()> {
