+++ bindings/wasm/src/account/wasm_account/account.rs
@@ -12,13 +12,13 @@ use identity_iota::account_storage::CekAlgorithm;
 use identity_iota::account_storage::EncryptedData;
 use identity_iota::account_storage::EncryptionAlgorithm;
 use identity_iota::account_storage::Storage;
+use identity_iota::client::Client;
 use identity_iota::core::OneOrMany;
 use identity_iota::credential::Credential;
 use identity_iota::credential::Presentation;
 use identity_iota::crypto::ProofOptions;
 use identity_iota::crypto::PublicKey;
 use identity_iota::did::verifiable::VerifiableProperties;
-use identity_iota::iota::Client;
 use identity_iota::iota_core::IotaDID;
 use identity_iota::iota_core::IotaDocument;
 use js_sys::Promise;
