+++ bindings/wasm/src/account/wasm_account/account_builder.rs
@@ -6,8 +6,8 @@ use std::rc::Rc;
 
 use identity_iota::account::AccountBuilder;
 use identity_iota::account::IdentitySetup;
-use identity_iota::iota::Client;
-use identity_iota::iota::ClientBuilder;
+use identity_iota::client::Client;
+use identity_iota::client::ClientBuilder;
 use identity_iota::iota_core::IotaDID;
 use js_sys::Promise;
 use wasm_bindgen::prelude::*;
