+++ bindings/wasm/src/account/wasm_account/update/create_method.rs
@@ -9,8 +9,8 @@ use identity_iota::account::CreateMethodBuilder;
 use identity_iota::account::IdentityUpdater;
 use identity_iota::account::MethodContent;
 use identity_iota::account::UpdateError::MissingRequiredField;
+use identity_iota::client::Client;
 use identity_iota::did::MethodScope;
-use identity_iota::iota::Client;
 use js_sys::Promise;
 use wasm_bindgen::prelude::*;
 use wasm_bindgen::JsCast;
