+++ bindings/wasm/src/account/wasm_account/update/create_service.rs
@@ -8,9 +8,9 @@ use std::rc::Rc;
 use identity_iota::account::CreateServiceBuilder;
 use identity_iota::account::IdentityUpdater;
 use identity_iota::account::UpdateError::MissingRequiredField;
+use identity_iota::client::Client;
 use identity_iota::core::Object;
 use identity_iota::did::ServiceEndpoint;
-use identity_iota::iota::Client;
 use js_sys::Promise;
 use wasm_bindgen::prelude::*;
 use wasm_bindgen::JsCast;
