+++ bindings/wasm/src/chain/document_history.rs
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::common::ArrayString;
-use identity_iota::iota::ChainHistory;
-use identity_iota::iota::DocumentHistory;
-use identity_iota::iota::ResolvedIotaDocument;
+use identity_iota::client::ChainHistory;
+use identity_iota::client::DocumentHistory;
+use identity_iota::client::ResolvedIotaDocument;
 use identity_iota::iota_core::DiffMessage;
 use wasm_bindgen::prelude::*;
 use wasm_bindgen::JsCast;
