+++ bindings/wasm/src/tangle/receipt.rs
@@ -1,7 +1,7 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use identity_iota::iota::Receipt;
+use identity_iota::client::Receipt;
 use wasm_bindgen::prelude::*;
 
 use crate::error::Result;
