+++ examples/account/config.rs
@@ -9,8 +9,8 @@ use identity_iota::account::AutoSave;
 use identity_iota::account::IdentitySetup;
 use identity_iota::account::Result;
 use identity_iota::account_storage::MemStore;
-use identity_iota::iota::ClientBuilder;
-use identity_iota::iota::ExplorerUrl;
+use identity_iota::client::ClientBuilder;
+use identity_iota::client::ExplorerUrl;
 use identity_iota::iota_core::IotaDID;
 use identity_iota::iota_core::Network;
