+++ examples/account/create_did.rs
@@ -9,7 +9,7 @@ use identity_iota::account::Account;
 use identity_iota::account::IdentitySetup;
 use identity_iota::account::Result;
 use identity_iota::account_storage::Stronghold;
-use identity_iota::iota::ExplorerUrl;
+use identity_iota::client::ExplorerUrl;
 use identity_iota::iota_core::IotaDID;
 
 #[tokio::main]
