+++ examples/account/lazy.rs
@@ -8,8 +8,8 @@ use identity_iota::account::Account;
 use identity_iota::account::IdentitySetup;
 use identity_iota::account::Result;
 use identity_iota::account_storage::Stronghold;
+use identity_iota::client::ExplorerUrl;
 use identity_iota::core::Url;
-use identity_iota::iota::ExplorerUrl;
 use identity_iota::iota_core::IotaDID;
 
 #[tokio::main]
