+++ examples/low-level-api/resolve_history.rs
@@ -6,6 +6,10 @@
 //!
 //! cargo run --example did_history
 
+use identity_iota::client::Client;
+use identity_iota::client::DocumentHistory;
+use identity_iota::client::Receipt;
+use identity_iota::client::Result;
 use identity_iota::core::json;
 use identity_iota::core::FromJson;
 use identity_iota::core::Timestamp;
@@ -13,10 +17,6 @@ use identity_iota::crypto::KeyPair;
 use identity_iota::did::MethodScope;
 use identity_iota::did::Service;
 use identity_iota::did::DID;
-use identity_iota::iota::Client;
-use identity_iota::iota::DocumentHistory;
-use identity_iota::iota::Receipt;
-use identity_iota::iota::Result;
 use identity_iota::iota_core::IotaDocument;
 use identity_iota::iota_core::IotaService;
 use identity_iota::iota_core::IotaVerificationMethod;
